Minor corrections
In particular, no action would allow the model to actually reach x =
length due to a mistake in the decision making + monitoring module.

+++ code/human_driver_model/two_module_model/automatic_model_checker.py
@@ -7,19 +7,27 @@
 # Email: francisco.eiras@cs.ox.ac.uk
 # 26-Apr-2018; Last revision: 27-Apr-2018
 
-import sys, os, subprocess
+import sys, os, subprocess, csv
 
 types = ['aggressive','average','cautious']
+res = {}
 
-v = 20
-v1 = 15
-x1_0 = 30
+cleaning_up = True
+properties_file = 'properties.pctl'
+v = 26
+v1 = 20
+x1_0 = 39
+
+num_properties = sum(1 for line in open(properties_file))
+with open(properties_file) as f1:
+    props = f1.readlines()
+props = [x.strip() for x in props]
 
 for driver_type in range(1,4):
 	print('------ %s driver ------'%types[driver_type-1])
 
 	filename = "gen_model_%s_%s_%s_%s"%(driver_type,v,v1,x1_0)
-	r_filename = "results_%s_%s_%s_%s"%(driver_type,v,v1,x1_0)
+	r_filename = "results_%s_%s_%s_%s_%s"%(driver_type,properties_file,v,v1,x1_0)
 
 	# Construct the file
 	print('Generating the model...')
@@ -31,14 +39,53 @@ for driver_type in range(1,4):
 	print('Obtaining the results...')
 
 	f = open("results/%s.txt"%r_filename, "r")
-	f.readline();
-	probability = f.readline();
-	probability = float(probability[:-1])
 
-	print('The probability of crashing for an %s driver is: %f.'%(types[driver_type-1],probability))
+	if num_properties == 1:
+		f.readline()
+		probability = f.readline()
+		f.close()
 
-	print('Cleaning up...')
+		probability = float(probability[:-1])
+		res[driver_type] = [props[0], probability]
+	else:
+		f.readline()
+		f.readline()
+		probability = f.readline()
 
-	os.system('rm results/%s.pm'%filename)
+		probability = float(probability[:-1])
+		res[driver_type] = [[props[0], probability]]
+
+		for i in range(1,num_properties):
+			f.readline()
+			f.readline()
+			f.readline()
+			probability = f.readline()
+
+			probability = float(probability[:-1])
+			res[driver_type].append([props[i], probability])
+
+	f.close()
+
+	if cleaning_up == True:
+		print('Cleaning up...')
+		os.system('rm results/%s.pm'%filename)
+		os.system('rm results/%s.txt'%r_filename)
+
+with open('results/%s_%s_%s_%s.csv'%(properties_file,v,v1,x1_0), 'w') as csvfile:
+    fieldnames = ['type_driver', 'property', 'probability']
+    writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
+
+    writer.writeheader()
+    for key,val in res.items():
+    	for propty in val:
+    		writer.writerow({'type_driver': key, 'property': propty[0], 'probability': propty[1]})
 
 print('Done.')
+
+
+
+
+
+
+
+
